Move all the gui classes in include/gdumper/gui/
